Fix links in docs & improve link in sidebar (#36)
* Docs: Fix links from index pages to child pages

- Need to be relative links, otherwise the base path gets lost

* Docs: Fix links to other docs pages

Went through the docs and corrected links that aren't working

* Fix link in blog post

* Docs: Add links to existing docs pages to sidebar

Lots of newer documentation pages weren't linked in the sidebar, plus the overview/index page per section was only linked in the header, but not sidebar. This way the docs are more accessible to the reader, and the breadth of docs is immediately more visible.